Refactor RenderTapOutputProfile to be in line with changes on master
